+++ events/messageUpdate.js
@@ -76,6 +76,9 @@ exports.run = async(client, oldMsg, msg) => {
 	if (cmd.help.botpermissions.every(perm => msg.guild.me.hasPermission(perm)) === false) return msg.channel.send(botnopermission);
 	if (cmd.conf.userpermissions.every(perm => msg.member.hasPermission(perm)) === false) return msg.channel.send(usernopermission);
 
+
+	if (cmd.conf.enabled === false) return msg.reply(lang.messageevent_commanddisabled);
+
 		cmd.run(client, msg, args, lang);
 		if (tableload.commanddel === 'true') {
 			msg.delete();
